+++ src/sentry/api/serializers/models/user_identity_config.py
@@ -12,11 +12,11 @@ from sentry.api.serializers import Serializer, register, serialize
 from sentry.api.serializers.models import ControlSiloOrganizationSerializer
 from sentry.auth.provider import Provider
 from sentry.exceptions import NotRegistered
+from sentry.hybridcloud.services.organization_mapping import organization_mapping_service
 from sentry.identity import is_login_provider
 from sentry.integrations.manager import default_manager as integrations
 from sentry.models.authidentity import AuthIdentity
 from sentry.models.identity import Identity
-from sentry.services.hybrid_cloud.organization_mapping import organization_mapping_service
 from social_auth.models import UserSocialAuth
 
 from . import user_social_auth
