+++ src/sentry/models/organization.py
@@ -26,11 +26,11 @@ from sentry.db.models.manager.base import BaseManager
 from sentry.db.models.outboxes import ReplicatedRegionModel
 from sentry.db.models.utils import slugify_instance
 from sentry.db.postgres.transactions import in_test_hide_transaction_boundary
+from sentry.hybridcloud.services.organization_mapping import organization_mapping_service
 from sentry.locks import locks
 from sentry.models.outbox import OutboxCategory
 from sentry.notifications.services import notifications_service
 from sentry.roles.manager import Role
-from sentry.services.hybrid_cloud.organization_mapping import organization_mapping_service
 from sentry.services.hybrid_cloud.user import RpcUser, RpcUserProfile
 from sentry.services.hybrid_cloud.user.service import user_service
 from sentry.types.organization import OrganizationAbsoluteUrlMixin
@@ -256,10 +256,10 @@ class Organization(ReplicatedRegionModel, OrganizationAbsoluteUrlMixin):
         return super().delete(**kwargs)
 
     def handle_async_replication(self, shard_identifier: int) -> None:
-        from sentry.services.hybrid_cloud.organization_mapping.serial import (
+        from sentry.hybridcloud.services.organization_mapping.serial import (
             update_organization_mapping_from_instance,
         )
-        from sentry.services.hybrid_cloud.organization_mapping.service import (
+        from sentry.hybridcloud.services.organization_mapping.service import (
             organization_mapping_service,
         )
         from sentry.types.region import get_local_region
