+++ src/sentry/notifications/notificationcontroller.py
@@ -7,6 +7,7 @@ from typing import Union
 from django.db.models import Q
 
 from sentry import features
+from sentry.hybridcloud.services.organization_mapping.serial import serialize_organization_mapping
 from sentry.integrations.types import (
     EXTERNAL_PROVIDERS_REVERSE_VALUES,
     PERSONAL_NOTIFICATION_PROVIDERS,
@@ -32,7 +33,6 @@ from sentry.notifications.types import (
     NotificationSettingEnum,
     NotificationSettingsOptionEnum,
 )
-from sentry.services.hybrid_cloud.organization_mapping.serial import serialize_organization_mapping
 from sentry.services.hybrid_cloud.user.model import RpcUser
 from sentry.types.actor import Actor, ActorType
