+++ src/sentry/services/hybrid_cloud/organization_mapping/__init__.py
@@ -1,2 +1,6 @@
-from .model import *  # noqa
-from .service import *  # noqa
+# TODO(hybridcloud) Remove once getsentry usage is updated
+from sentry.hybridcloud.services.organization_mapping.model import RpcOrganizationMapping
+from sentry.hybridcloud.services.organization_mapping.service import organization_mapping_service
+from sentry.silo.base import SiloMode
+
+__all__ = ("organization_mapping_service", "RpcOrganizationMapping", "SiloMode")
