+++ src/sentry/services/hybrid_cloud/user/impl.py
@@ -17,6 +17,8 @@ from sentry.api.serializers import (
 from sentry.api.serializers.base import Serializer, serialize
 from sentry.db.models.query import in_iexact
 from sentry.hybridcloud.rpc.filter_query import FilterQueryDatabaseImpl, OpaqueSerializedResponse
+from sentry.hybridcloud.services.organization_mapping.model import RpcOrganizationMapping
+from sentry.hybridcloud.services.organization_mapping.serial import serialize_organization_mapping
 from sentry.models.authidentity import AuthIdentity
 from sentry.models.avatars import UserAvatar
 from sentry.models.organization import OrganizationStatus
@@ -25,8 +27,6 @@ from sentry.models.organizationmembermapping import OrganizationMemberMapping
 from sentry.models.user import User
 from sentry.models.useremail import UserEmail
 from sentry.services.hybrid_cloud.auth import AuthenticationContext
-from sentry.services.hybrid_cloud.organization_mapping.model import RpcOrganizationMapping
-from sentry.services.hybrid_cloud.organization_mapping.serial import serialize_organization_mapping
 from sentry.services.hybrid_cloud.user import (
     RpcAvatar,
     RpcUser,
