+++ tests/sentry/features/test_flagpole_context.py
@@ -9,9 +9,9 @@ from sentry.features.flagpole_context import (
     project_context_transformer,
     user_context_transformer,
 )
+from sentry.hybridcloud.services.organization_mapping import organization_mapping_service
 from sentry.models.useremail import UserEmail
 from sentry.services.hybrid_cloud.organization import organization_service
-from sentry.services.hybrid_cloud.organization_mapping import organization_mapping_service
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import control_silo_test
